Revise script in general

+++ install.sh
@@ -26,52 +26,23 @@ target_dir="/opt/salomon"
 clean_install=0
 git_clone=".git snippets wiki README.md"
 keep_directory=0
-target="${cl_yl}${target_dir}${cl_n}"
 yesno="${cl_yl}Y${cl_n}/${cl_yl}N${cl_n}"
 proceed="Do you wish to proceed"
 
 available="everyone"
+is_root=0
 dirmod=775
 filemod=664
 execmod=$dirmod
 
-perform() {
-    echo
-    echo -e "${cl_lc}Salomon install/uninstall script${cl_n}"
-    echo
-    if [ "$script_mode" = "install" ]; then
-        echo -e "Installing Salomon is ${cl_yl}optional${cl_n} and not" \
-                "mandatory in order to use it. Further"
-        echo    "information can be found inside the documentation file for" \
-                "this script."
-        echo
-    fi
-    confirm "This will $script_action Salomon. $proceed ($yesno)? \c"
-    if [ $choice -eq 0 ]; then
-        echo
-        echo -e "${cl_lr}Canceled${cl_n} on user request."
-        echo
-        exit
-    fi
-    echo
-    if [ "$script_mode" = "install" ]; then
-        echo -e "You can make Salomon available either for all users on" \
-                "this machine or only"
-        echo -e "for root. \c"
-        confirm "Do you want to make it available for all users ($yesno)? \c"
-        if [ $choice -eq 0 ]; then
-            available="rootonly"
-        fi
-        echo
-    fi
-}
-
 set_permissions() {
-    chown -R root:root $target_dir
-    if [ "$available" = "rootonly" ]; then
-        dirmod=770
-        filemod=660
-        execmod=$dirmod
+    if [ $is_root -eq 1 ]; then
+        chown -R root:root $target_dir
+        if [ "$available" = "rootonly" ]; then
+            dirmod=770
+            filemod=660
+            execmod=$dirmod
+        fi
     fi
     find $target_dir > $temp_file
     while read line; do
@@ -136,10 +107,8 @@ else
     usage "Invalid argument" "$1"
 fi
 
-if [ "$(whoami)" != "root" ]; then
-    error="${cl_lr}error:${cl_n}"
-    echo -e "$error Superuser privileges are required."
-    exit 1
+if [ "$(whoami)" = "root" ]; then
+    is_root=1
 fi
 
 grep "\/usr\/local\/bin" <<< $PATH &>/dev/null
@@ -149,16 +118,66 @@ else
     symlink_sh="/usr/bin"
 fi
 
-perform $script_mode
+echo
+echo -e "${cl_lc}Salomon install/uninstall script${cl_n}"
+echo
+
+if [ "$script_mode" = "install" ]; then
+    echo -e "Installing Salomon is ${cl_yl}optional${cl_n} and not" \
+            "mandatory in order to use it. Further"
+    echo    "information can be found inside the documentation file for" \
+            "this script."
+    echo
+fi
+
+if [ $is_root -eq 0 ]; then
+    if [ "$script_mode" = "install" ]; then
+        target_dir="$HOME/salomon"
+        echo    "You are not running this script as superuser. Due to this," \
+                "a system-wide"
+        echo    "installation cannot be performed. Instead, Salomon will be" \
+                "installed into"
+        echo -e "your home directory '${cl_yl}$HOME${cl_n}'."
+        echo
+    fi
+fi
+target="${cl_yl}${target_dir}${cl_n}"
+
+confirm "This will $script_action Salomon. $proceed ($yesno)? \c"
+if [ $choice -eq 0 ]; then
+    echo
+    echo -e "${cl_lr}Canceled${cl_n} on user request."
+    echo
+    exit
+fi
+echo
+
+if [ "$script_mode" = "install" ]; then
+    echo -e "The current target directory is '$target'."
+    echo
+    echo -e "You can also include the version number in the installation" \
+            "path name (with"
+    echo -e "this version the path name would be" \
+            "'${target}${cl_yl}-$version${cl_n}' then)."
+    confirm "Do you want to include the version in the path ($yesno)? \c"
+    if [ $choice -eq 1 ]; then
+        target_dir="${target_dir}-$version"
+        target="${cl_yl}${target_dir}${cl_n}"
+    fi
+    echo
+fi
+
 if [ $script_mode = "install" ]; then
     if [ -d "$target_dir" ]; then
         if [ ! "$script_dir" = "$target_dir" ]; then
-            echo -e "The target directory '$target' already exists. You can" \
-                    "perform a clean"
-            echo -e "installation which will delete the directory and" \
-                    "reinstall the original files."
-            echo -e "Notice that all user-defined configs and settings will" \
-                    "also be deleted then."
+            echo -e "The target directory '$target' already exists."
+            echo
+            echo -e "You can perform a clean installation which will delete" \
+                    "the directory and"
+            echo -e "reinstall the original files. Notice that all" \
+                    "user-defined configuration"
+            echo -e "as well as color and filter files will also be deleted" \
+                    "then."
             confirm "Do you want to perform a clean installation ($yesno)? \c"
             clean_install=$choice
             echo
@@ -173,6 +192,20 @@ if [ $script_mode = "install" ]; then
                      "files) is not possible."
                 echo
             fi
+        fi
+    fi
+
+    if [ $is_root -eq 1 ]; then
+        if [ "$script_mode" = "install" ]; then
+            echo -e "You can make Salomon available either for all users on" \
+                    "this machine or only"
+            echo -e "for root. \c"
+            confirm \
+              "Do you want to make it available for all users ($yesno)? \c"
+            if [ $choice -eq 0 ]; then
+                available="rootonly"
+            fi
+            echo
         fi
     fi
 
@@ -203,44 +236,52 @@ if [ $script_mode = "install" ]; then
         rm -fR $target_dir/$git_clone &>/dev/null
     done
 
-    echo -e "Setting permissions for installation directory... \c"
-    if [ $available = "rootonly" ]; then
-        echo -e "${cl_lb}(root only)${cl_n}"
-    else
-        echo -e "${cl_lb}(everyone)${cl_n}"
-    fi
-    set_permissions
+    if [ $is_root -eq 1 ]; then
+        echo -e "Setting permissions for installation directory... \c"
+          if [ $available = "rootonly" ]; then
+            echo -e "${cl_lb}(root only)${cl_n}"
+        else
+            echo -e "${cl_lb}(everyone)${cl_n}"
+        fi
+        set_permissions
 
-    if [ $clean_install -eq 1 ]; then
+        if [ $clean_install -eq 1 ]; then
+            if [ -f ${symlink_sh}/salomon ]; then
+                echo "Removing existing symbolic link for main script..."
+                rm -f ${symlink_sh}/salomon &>/dev/null
+            fi
+        fi
+
+        echo -e "Creating symbolic link for main script... \c"
         if [ -f ${symlink_sh}/salomon ]; then
-            echo "Removing existing symbolic link for main script..."
-            rm -f ${symlink_sh}/salomon &>/dev/null
+            echo -e "${cl_lb}(already exists)${cl_n}"
+        else
+            ln -s ${target_dir}/salomon.sh ${symlink_sh}/salomon &>/dev/null
+            echo
         fi
     fi
-
-    echo -e "Creating symbolic link for main script... \c"
-    if [ -f ${symlink_sh}/salomon ]; then
-        echo -e "${cl_lb}(already exists)${cl_n}"
-    else
-        ln -s ${target_dir}/salomon.sh ${symlink_sh}/salomon &>/dev/null
-        echo
-    fi
+    echo
 else  # uninstall
+    target_dir=$script_dir
+    target="${cl_yl}${target_dir}${cl_n}"
     echo -e "Removing the installation directory '$target' will also" \
             "delete all"
-    echo -e "user-defined configs and settings. \c"
+    echo -e "all user-defined configuration as well as color and filter" \
+            "files."
     confirm "Do you want to remove it ($yesno)? \c"
     if [ $choice -eq 0 ]; then
         keep_directory=1
     fi
     echo
 
-    echo -e "Removing symbolic link for main script... \c"
-    if [ -f ${symlink_sh}/salomon ]; then
-        rm -f ${symlink_sh}/salomon &>/dev/null
-        echo
-    else
-        echo -e "${cl_lb}(does not exist)${cl_n}"
+    if [ $is_root -eq 1 ]; then
+        echo -e "Removing symbolic link for main script... \c"
+        if [ -f ${symlink_sh}/salomon ]; then
+            rm -f ${symlink_sh}/salomon &>/dev/null
+            echo
+        else
+            echo -e "${cl_lb}(does not exist)${cl_n}"
+        fi
     fi
 
     echo -e "Removing installation directory '${target}'... \c"
@@ -249,17 +290,24 @@ else  # uninstall
     else
         if [ -d $target_dir ]; then
             rm -fR $target_dir &>/dev/null
-            echo
         else
             echo -e "${cl_lb}(does not exist)${cl_n}"
         fi
     fi
+    echo
 fi
-echo
+
 echo -e "Salomon has been ${script_mode}ed."
-if [ $script_mode = "install" ]; then
-    echo -e "You can now directly run the '${cl_yl}salomon${cl_n}' command" \
-            "in order to use it."
+if [ $is_root -eq 1 ]; then
+    if [ $script_mode = "install" ]; then
+        echo -e "You can now directly run the '${cl_yl}salomon${cl_n}'" \
+                "command in order to use it."
+    fi
+else
+    if [ $script_mode = "install" ]; then
+        cd $target_dir
+        ln -s salomon.sh salomon &>/dev/null
+    fi
 fi
 echo
